+++ backend/src/main/java/ProyectoFinalTienda/TiendaVideojuegos/dtos/responses/InventarioItemResponse.java
@@ -3,6 +3,8 @@ package ProyectoFinalTienda.TiendaVideojuegos.dtos.responses;
 import ProyectoFinalTienda.TiendaVideojuegos.model.enums.Plataformas;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class InventarioItemResponse {
     private int inventarioId;
     private VideojuegoResponse videojuego;
     private Plataformas plataforma;
-    private double precioDiario;
+    private BigDecimal precioDiario;
     private int stockTotal;
     private int stockDisponible;
 //    private int stockAlquilado;
